添加常用分类

+++ MyProject/Classes/UISubviews/Controller/UICollectionView/PhotoAlbumSelectImage/YHGetPhotoAlbumImageViewController.m
@@ -150,7 +150,7 @@ static NSString *const imageID = @"YHSelectImageCollectionViewCell";
         }
         [self.imageCollectionView reloadData];
     }
-    PPLog(@"selectImageAmountInteger -  %ld",self.selectImageAmountInteger);
+    PPLog(@"selectImageAmountInteger -  %ld",(long)self.selectImageAmountInteger);
     if (self.selectImageAmountInteger > 0) {
         [self.bottomView changeColor:true];
     } else {
